feat: added genre cards
