+++ apps/web/app/root.jsx
@@ -4,6 +4,7 @@ import i18next from "./i18next.server";
 import stylesheet from "./tailwind.css?url";
 import { json } from "@remix-run/node";
 import { useLoaderData } from "@remix-run/react";
+import Header from "./components/header.jsx";
 
 export const meta = () => [{ title: "Radio Pronto!" }];
 export const links = () => [{ rel: "stylesheet", href: stylesheet }];
@@ -23,13 +24,15 @@ export function Layout({ children }) {
         <Links />
       </head>
       <body className="bg-gray-100">
-        {children}
+
         <Scripts />
         <script
           defer
           data-domain="radiopronto.net"
           src="https://plausible.io/js/script.js"
         ></script>
+            <Header />
+            {children}
       </body>
     </html>
   );
