+++ apps/web/app/tailwind.css
@@ -9,3 +9,4 @@
     background-repeat: no-repeat;
   }
   
+
